webui: move pasword form component into a separate file
